+++ plugin-src/translators/text/font/custom/translateCustomFont.ts
@@ -1,15 +1,13 @@
-import { getCustomFontId, translateFontVariantId } from '@plugin/translators/text/font/custom';
+import { translateFontVariantId } from '@plugin/translators/text/font/custom';
 
 import { TextTypography } from '@ui/lib/types/shapes/textShape';
 
 export const translateCustomFont = (
-  fontName: FontName,
+  fontName: FontName | undefined,
   fontWeight: string
 ): Pick<TextTypography, 'fontId' | 'fontVariantId' | 'fontWeight'> | undefined => {
-  const customFontId = getCustomFontId(fontName);
-
   return {
-    fontId: customFontId ? `custom-${customFontId}` : '',
+    fontId: '',
     fontVariantId: translateFontVariantId(fontName, fontWeight),
     fontWeight
   };