+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # penpot-exporter
 
+## 0.13.1
+
+### Patch Changes
+
+- [#246](https://github.com/penpot/penpot-exporter-figma-plugin/pull/246)
+  [`05ec8ea`](https://github.com/penpot/penpot-exporter-figma-plugin/commit/05ec8ea2bc6f381a670beb84f16fb18713ef42f1)
+  Thanks [@Cenadros](https://github.com/Cenadros)! - Fixes undefined fontnames while fetching fonts
+
+## 0.13.0
+
+### Minor Changes
+
+- [#243](https://github.com/penpot/penpot-exporter-figma-plugin/pull/243)
+  [`b44cbf2`](https://github.com/penpot/penpot-exporter-figma-plugin/commit/b44cbf2052e18d5d8a77447af48a77e2c3f54b9a)
+  Thanks [@Cenadros](https://github.com/Cenadros)! - Remove custom font form since penpot now
+  matches the fonts by fontName and fontFamily
+
+### Patch Changes
+
+- [#239](https://github.com/penpot/penpot-exporter-figma-plugin/pull/239)
+  [`570aeee`](https://github.com/penpot/penpot-exporter-figma-plugin/commit/570aeee7b0a5298ca7ed4e99aaaf216ecd621cc8)
+  Thanks [@Cenadros](https://github.com/Cenadros)! - Minor updates
+
 ## 0.12.1
 
 ### Patch Changes